# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bissi / Committee / Chit Fund Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List all branches
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Branch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    get:
      operationId: getBranch
      tags: [branches]
      summary: Get branch by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    patch:
      operationId: updateBranch
      tags: [branches]
      summary: Update a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchUpdate"
      responses:
        "200":
          description: Branch updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete a branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Customers
...
  /collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List collections/payments
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: collectorId
          in: query
          schema:
            type: integer
        - name: committeeId
          in: query
          schema:
            type: integer
        - name: loanId
          in: query
          schema:
            type: integer
        - name: branchId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of collections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CollectionListResponse"
    post:
      operationId: createCollection
      tags: [collections]
      summary: Record a collection/payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionInput"
      responses:
        "201":
          description: Collection recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
  /collections/{id}:
    get:
      operationId: getCollection
      tags: [collections]
      summary: Get collection by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
...
components:
  schemas:
...
    Branch:
      type: object
      required: [id, name, code, city, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        city:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        managerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        totalCustomers:
          type: integer
        totalCollectors:
          type: integer
        createdAt:
          type: string
    BranchInput:
      type: object
      required: [name, code, city]
      properties:
        name:
          type: string
        code:
          type: string
        city:
          type: string
        address:
          type: string
        phone:
          type: string
        managerName:
          type: string
        status:
          type: string
          enum: [active, inactive]
    BranchUpdate:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        city:
          type: string
        address:
          type: string
        phone:
          type: string
        managerName:
          type: string
        status:
          type: string
          enum: [active, inactive]
...
    Collection:
      type: object
      required: [id, customerId, amount, paymentMode, collectedAt, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        customerMobile:
          type: ["string", "null"]
        collectorId:
          type: ["integer", "null"]
        collectorName:
          type: ["string", "null"]
        branchId:
          type: ["integer", "null"]
        committeeId:
          type: ["integer", "null"]
        committeeName:
          type: ["string", "null"]
        loanId:
          type: ["integer", "null"]
        amount:
          type: number
        paymentMode:
          type: string
          enum: [cash, upi, bank, card]
        receiptNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        collectedAt:
          type: string
        createdAt:
          type: string
        billingName:
          type: ["string", "null"]
        billingPhone:
          type: ["string", "null"]
        billingAddress:
          type: ["string", "null"]
        billingGstin:
          type: ["string", "null"]
    CollectionInput:
      type: object
      required: [customerId, amount, paymentMode]
      properties:
        customerId:
          type: integer
        collectorId:
          type: integer
        committeeId:
          type: integer
        loanId:
          type: integer
        amount:
          type: number
        paymentMode:
          type: string
          enum: [cash, upi, bank, card]
        notes:
          type: string
        collectedAt:
          type: string
        billingName:
          type: string
        billingPhone:
          type: string
        billingAddress:
          type: string
        billingGstin:
          type: string
    CollectionListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Collection"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...