# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bissi / Committee / Chit Fund Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CollectionPaymentMode } from './collectionPaymentMode';

export interface Collection {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerMobile?: string | null;
  /** @nullable */
  collectorId?: number | null;
  /** @nullable */
  collectorName?: string | null;
  /** @nullable */
  branchId?: number | null;
  /** @nullable */
  committeeId?: number | null;
  /** @nullable */
  committeeName?: string | null;
  /** @nullable */
  loanId?: number | null;
  amount: number;
  paymentMode: CollectionPaymentMode;
  /** @nullable */
  receiptNumber?: string | null;
  /** @nullable */
  notes?: string | null;
  collectedAt: string;
  createdAt: string;
  /** @nullable */
  billingName?: string | null;
  /** @nullable */
  billingPhone?: string | null;
  /** @nullable */
  billingAddress?: string | null;
  /** @nullable */
  billingGstin?: string | null;
}