# lib/api-zod/src/generated/types/collectionInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Bissi / Committee / Chit Fund Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CollectionInputPaymentMode } from './collectionInputPaymentMode';

export interface CollectionInput {
  customerId: number;
  collectorId?: number;
  committeeId?: number;
  loanId?: number;
  amount: number;
  paymentMode: CollectionInputPaymentMode;
  notes?: string;
  collectedAt?: string;
  billingName?: string;
  billingPhone?: string;
  billingAddress?: string;
  billingGstin?: string;
}